[prog] moved write file code to seeder file
